+++ Genesis/Changes.md
@@ -8,6 +8,8 @@ value. Added appropriate text scaling to the method.
     - A simple class which allows you to easily create a parable-like projectile motion for your double value
 - Direction1D
     - A very simple enumeration for one dimensional direction (forward / backward)
+- BezierFunction
+    - Allows you to create a smooth x-y function based on specified points
 ### Deprecations
 - Direction2D.isPositiveDirection
     - Replaced with .sign and Direction1D